refactor(uttaksplan): flytt regelkatalog-OMRÅDE-konstantar inn i stories-filene
Dei *_OMRÅDE-konstantane som berre brukast frå Storybook-katalogen
(metadata + samla regelliste for ein gitt regelgruppe) ligg no inne
i sjølve stories-filene. Regelfilene eksponerer berre reglane sjølve
+ runtime-helpers, og treng ikkje lenger importere område-typane.

Valideringsområda (SAMTIDIG_UTTAK_OMRÅDE m.fl.) blir framleis brukt
frå valider.ts og ligg difor uendra.

Rein refaktor — ingen funksjonalitet er endra.
