+++ packages/uttaksplan/src/regler/Feltregler.stories.tsx
@@ -3,9 +3,37 @@ import { useIntl } from 'react-intl';
 
 import { Kolonne, MeldingSitat, RegelIdBadge, RegelkatalogSide } from './RegelkatalogSide';
 import { Feltregelområde } from './types';
-import { HVA_VIL_DU_GJØRE_OMRÅDE } from './felt/hvaVilDuGjøre';
-import { SAMTIDIG_UTTAKSPROSENT_OMRÅDE } from './felt/samtidigUttaksprosent';
-import { STILLINGSPROSENT_OMRÅDE } from './felt/stillingsprosent';
+import { HVA_VIL_DU_GJØRE_REGLER } from './felt/hvaVilDuGjøre';
+import { SAMTIDIG_UTTAKSPROSENT_REGLER } from './felt/samtidigUttaksprosent';
+import { STILLINGSPROSENT_REGLER } from './felt/stillingsprosent';
+
+const STILLINGSPROSENT_OMRÅDE: Feltregelområde = {
+    id: 'stillingsprosent',
+    område: 'Stillingsprosent',
+    beskrivelse:
+        'Hvor stor andel brukeren skal jobbe når arbeid kombineres med uttak. Reglene sikrer at verdien er ' +
+        'et gyldig tall mellom 0 og 100 %, og at summen med samtidig uttak ikke overstiger 100 %.',
+    regler: STILLINGSPROSENT_REGLER,
+};
+
+const SAMTIDIG_UTTAKSPROSENT_OMRÅDE: Feltregelområde = {
+    id: 'samtidigUttaksprosent',
+    område: 'Samtidig uttaksprosent',
+    beskrivelse:
+        'Hvor stor andel av kvoten brukeren tar ut samtidig med den andre forelderen. Reglene sikrer at verdien ' +
+        'er et gyldig tall mellom 0 og 100 %, og at summen med stillingsprosent ikke overstiger 100 %.',
+    regler: SAMTIDIG_UTTAKSPROSENT_REGLER,
+};
+
+const HVA_VIL_DU_GJØRE_OMRÅDE: Feltregelområde = {
+    id: 'hvaVilDuGjøre',
+    område: 'Hva vil du gjøre?',
+    beskrivelse:
+        'Når brukeren velger hva som skal legges til i uttaksplanen (ferie, utsettelse, pause, opphold eller ' +
+        'periode), sjekker vi at valget er gyldig i forhold til når perioden ligger og hvilke rettigheter ' +
+        'brukeren har.',
+    regler: HVA_VIL_DU_GJØRE_REGLER,
+};
 
 const ALLE_FELTREGLER: readonly Feltregelområde[] = [
     STILLINGSPROSENT_OMRÅDE,
