+++ packages/uttaksplan/src/regler/kvotetype/kvoteRegler.ts
@@ -3,7 +3,7 @@ import type { BrukerRolleSak_fpoversikt, Familiesituasjon, KontoTypeUttak, Retti
 import { useUttaksplanData } from '../../context/UttaksplanDataContext';
 import { UttaksperiodeValidatorer } from '../../utils/UttaksperiodeValidatorer';
 import { Periode } from '../types';
-import { Kvoteregel, Kvoteområde } from './types';
+import { Kvoteregel } from './types';
 
 export type KvoteKontekst = {
     søker: BrukerRolleSak_fpoversikt;
@@ -271,24 +271,6 @@ export const ALLE_KVOTE_REGLER: ReadonlyArray<Kvoteregel<KvoteKontekst>> = [
     ...MOR_KVOTE_REGLER,
     ...FAR_MEDMOR_KVOTE_REGLER,
 ];
-
-export const MOR_KVOTE_OMRÅDE: Kvoteområde = {
-    id: 'morKvoter',
-    område: 'Hvilke kvotetyper er gyldige å velge for mor?',
-    beskrivelse:
-        'Når mor er valgt som forelder for en periode, må kvotetypen (stønadskonto) være gyldig. ' +
-        'Reglene under bestemmer hvilke kvoter mor får tilbudt i nedtrekkslisten.',
-    regler: MOR_KVOTE_REGLER,
-};
-
-export const FAR_MEDMOR_KVOTE_OMRÅDE: Kvoteområde = {
-    id: 'farMedmorKvoter',
-    område: 'Hvilke kvotetyper er gyldige å velge for far/medmor?',
-    beskrivelse:
-        'Når far/medmor er valgt som forelder for en periode, må kvotetypen (stønadskonto) være gyldig. ' +
-        'Reglene under bestemmer hvilke kvoter far/medmor får tilbudt i nedtrekkslisten.',
-    regler: FAR_MEDMOR_KVOTE_REGLER,
-};
 
 /* ---------- Helpers og hook ---------- */
